requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai>=1.0.0
apscheduler==3.10.4
aiofiles==23.2.1
